with_first_found os dependent variables and tasks
Adopt the `with_first_found` method when including platform dependent
variables and tasks. This allows us to change details in specific OS
versions as neeed without extra when conditions. The platform dependent
files do not cascade, each one is independent, so some duplication is
expected.

The search order (with examples) is:

1. Distro and full version (ubuntu16.04, centos7.3.1611, debian8.7)
2. Distro and major version (ubuntu16, centos7, debian9)
3: Distro only (ubuntu, centos, debian)
4: OS Family (debian, redhat)