Select installation tasks by os type
Instead of selecting installation tasks by package manager name, select
the tasks by os type, using with_first_found, to select from most
specific to the most general. This allows us to make variations based on
os versions, and to deal with the fact that we use the same tasks for
dnf based systems as yum.

Consolidate some of the extra files for the kmod vs dkms installs.